Make it possible to save new files in a temporary file

+++ cgfs.py
@@ -1,10 +1,14 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from __future__ import division, print_function, absolute_import
+from os import getenv
+
+import uuid
+from pathlib import Path
 
 import logging
-from os import O_EXCL, O_CREAT, O_TRUNC
+import os
+from os import O_EXCL, O_CREAT, O_TRUNC, path
 from enum import IntEnum
 from stat import S_IFDIR, S_IFREG
 from time import time
@@ -14,6 +18,7 @@ from errno import (
 from getpass import getpass
 from argparse import ArgumentParser
 
+import tempfile
 from fuse import FUSE, Operations, FuseOSError, LoggingMixIn
 from cgapi import CGAPI, APICodes, CGAPIException
 
@@ -38,7 +43,7 @@ class DirTypes(IntEnum):
 
 class BaseFile():
     def __init__(self, data, name=None):
-        self.id = data['id']
+        self.id = data.get('id', None)
         self.name = name if name is not None else data['name']
         self.stat = None
 
@@ -49,6 +54,8 @@ class BaseFile():
                 'st_atime': time(),
                 'st_mtime': time(),
                 'st_ctime': time(),
+                'st_uid': os.getuid(),
+                'st_gid': os.getegid(),
             }
 
             if submission is not None and path is not None:
@@ -84,13 +91,17 @@ class Directory(BaseFile):
         return self.stat
 
     def insert(self, file):
+        if self.stat is None:
+            self.getattr()
+
         self.children[file.name] = file
         self.stat['st_nlink'] += 1
 
     def get(self, filename):
-        if filename not in self.children:
+        try:
+            return self.children[filename]
+        except KeyError:
             raise FuseOSError(ENOENT)
-        return self.children[filename]
 
     def pop(self, filename):
         try:
@@ -102,15 +113,120 @@ class Directory(BaseFile):
         return file
 
     def read(self):
-        return ['.', '..'] + [x for x in self.children]
+        res = list(self.children)
+        res.extend(['.', '..'])
+        return res
+
+
+class TempDirectory(Directory):
+    def __init__(self, *args, **kwargs):
+        super(TempDirectory, self).__init__(*args, **kwargs)
+        self.stat = {
+            'st_size': 0,
+            'st_atime': time(),
+            'st_ctime': time(),
+            'st_size': 0,
+            'st_mtime': time(),
+            'st_mode': S_IFDIR | 0o770,
+            'st_nlink': 2,
+        }
+
+
+class TempFile:
+    def __init__(self, name, tmpdir):
+        self._tmpdir = tmpdir
+        self.name = name
+
+        self.__handle = None
+
+        # Create a new temporary file
+        self._filename = str(uuid.uuid4())
+
+        while path.exists(path.join(path.join(self._tmpdir, self._filename))):
+            self._filename = str(uuid.uuid4())
+
+        self.full_path = path.join(self._tmpdir, self._filename)
+        Path(path.join(self.full_path)).touch()
+
+    @property
+    def stat(self):
+        st = os.lstat(self.full_path)
+        stat = {
+            key: getattr(st, key)
+            for key in (
+                'st_atime', 'st_ctime', 'st_gid', 'st_mode', 'st_mtime',
+                'st_nlink', 'st_size', 'st_uid'
+            )
+        }
+        stat['st_mode'] = S_IFREG | 0o770
+        return stat
+
+    @property
+    def _handle(self):
+        if self.__handle is None:
+            self.__handle = open(self.full_path, 'r+b')
+            self.__handle.seek(0)
+        return self.__handle
+
+    def getattr(self):
+        return self.stat
+
+    def setattr(self, key, value):
+        raise ValueError
+
+    def utimens(self, atime, mtime):
+        os.utime(self.full_path, (atime, mtime))
+
+    def open(self, *args):
+        pass
+
+    def read(self, offset, size):
+        self._handle.seek(offset)
+        return self._handle.read(size)
+
+    def write(self, data, offset):
+        handle = self._handle
+        handle.seek(offset)
+        res = handle.write(data)
+        handle.flush()
+        return res
+
+    def release(self):
+        if self.__handle is not None:
+            self.__handle.flush()
+            self.__handle.close()
+        self.__handle = None
+
+    def flush(self):
+        self._handle.flush()
+
+    def unlink(self):
+        self.release()
+        os.unlink(self.full_path)
+        self._filename = None
+
+    def truncate(self, length):
+        self._handle.seek(0)
+        self._handle.truncate(length)
 
 
 class File(BaseFile):
     def __init__(self, data, name=None):
         super(File, self).__init__(data, name)
 
-        self.data = None
+        self._data = None
         self.dirty = False
+
+    @property
+    def data(self):
+        if self._data is None:
+            self._data = cgapi.get_file(self.id)
+            self.stat['st_size'] = len(self._data)
+        return self._data
+
+    @data.setter
+    def data(self, data):
+        self._data = data
 
     def getattr(self, submission=None, path=None):
         if self.stat is None:
@@ -122,16 +238,17 @@ class File(BaseFile):
         return self.stat
 
     def open(self, buf):
-        self.data = buf
-        self.stat['st_size'] = len(buf)
+        self._data = buf
         self.stat['st_atime'] = time()
 
     def flush(self):
         if not self.dirty:
             return
 
+        assert self._data is not None
+
         try:
-            res = cgapi.patch_file(self.id, self.data)
+            res = cgapi.patch_file(self.id, self._data)
         except CGAPIException as e:
             self.data = None
             self.dirty = False
@@ -140,18 +257,16 @@ class File(BaseFile):
         self.dirty = False
         return res
 
-    def read(self, offset, length):  # pragma: no cover
-        pass
-
     def release(self):
         self.data = None
 
     def truncate(self, length):
-        data = self.data
-        if length <= self.stat['st_size']:
-            self.data = data[:length]
+        if length == 0:
+            self._data = bytes('', 'utf8')
+        elif length <= self.stat['st_size']:
+            self._data = self.data[:length]
         else:
-            self.data = data + bytes(
+            self._data = self.data + bytes(
                 '\0' * (length - self.stat['st_size']), 'utf8'
             )
         self.stat['st_size'] = length
@@ -160,14 +275,20 @@ class File(BaseFile):
         self.dirty = True
 
     def write(self, data, offset):
-        if offset > len(self.data):
-            self.data += bytes(offset - len(self.data))
-        if len(self.data) - offset - len(data) > 0:
+        if offset > self.stat['st_size']:
+            self._data += bytes(offset - len(self.data))
+
+        if self.stat['st_size'] - offset - len(data) > 0:
+            # Write in between
             second_offset = offset + len(data)
-            self.data = self.data[:offset] + data + self.data[second_offset:]
+            old_d = self.data
+            self._data = old_d[:offset] + data + old_d[second_offset:]
+        elif offset == 0:
+            self._data = data
         else:
-            self.data = self.data[:offset] + data
-        self.stat['st_size'] = len(self.data)
+            self._data = self.data[:offset] + data
+
+        self.stat['st_size'] = len(self._data)
         self.stat['st_atime'] = time()
         self.stat['st_mtime'] = time()
         self.dirty = True
@@ -175,10 +296,14 @@ class File(BaseFile):
 
 
 class CGFS(LoggingMixIn, Operations):
-    def __init__(self, latest_only):
+    def __init__(self, latest_only, fixed=False, tmpdir=None):
         self.latest_only = latest_only
+        self.fixed = fixed
         self.files = {}
         self.fd = 0
+        self._open_files = {}
+
+        self._tmpdir = tmpdir
 
         self.files = Directory(
             {
@@ -186,8 +311,10 @@ class CGFS(LoggingMixIn, Operations):
                 'name': 'root'
             }, type=DirTypes.FSROOT
         )
+
         self.files.getattr()
         self.load_courses()
+        print('Mounted')
 
     def load_courses(self):
         for course in cgapi.get_courses():
@@ -265,19 +392,23 @@ class CGFS(LoggingMixIn, Operations):
             if part == '':  # pragma: no cover
                 continue
 
-            if not isinstance(file, Directory):  # pragma: no cover
-                raise FuseOSError(ENOTDIR)
+            try:
+                if not file.children:
+                    if file.type == DirTypes.ASSIGNMENT:
+                        self.load_submissions(file)
+                    elif file.type == DirTypes.SUBMISSION:
+                        self.load_submission_files(file)
+            except AttributeError:  # pragma: no cover
+                if not isinstance(file, Directory):
+                    raise FuseOSError(ENOTDIR)
+                raise
 
-            if len(file.children) == 0:
-                if file.type == DirTypes.ASSIGNMENT:
-                    self.load_submissions(file)
-                elif file.type == DirTypes.SUBMISSION:
-                    self.load_submission_files(file)
-
-            file = file.get(part)
+            if part not in file.children:
+                raise FuseOSError(ENOENT)
+            file = file.children[part]
 
         if expect_type is not None:
-            if expect_type is File and not isinstance(file, File):
+            if expect_type is File and not isinstance(file, (File, TempFile)):
                 raise FuseOSError(EISDIR)
             elif expect_type is Directory and not isinstance(file, Directory):
                 raise FuseOSError(ENOTDIR)
@@ -307,19 +438,25 @@ class CGFS(LoggingMixIn, Operations):
             submission = self.get_submission(path)
             query_path = submission.tld + '/' + '/'.join(parts[3:])
 
-            try:
-                fdata = cgapi.create_file(submission.id, query_path)
-            except CGAPIException as e:
-                handle_cgapi_exception(e)
+            if self.fixed:
+                file = TempFile(fname, self._tmpdir)
+            else:
+                try:
+                    fdata = cgapi.create_file(submission.id, query_path)
+                except CGAPIException as e:
+                    handle_cgapi_exception(e)
 
-            file = File(fdata, name=fname)
-            file.setattr('st_size', fdata['size'])
-            file.setattr('st_mtime', fdata['modification_date'])
+                file = File(fdata, name=fname)
+                file.setattr('st_size', fdata['size'])
+                file.setattr('st_mtime', fdata['modification_date'])
+
             parent.insert(file)
 
         file.open(bytes('', 'utf8'))
 
         self.fd += 1
+        self._open_files[self.fd] = file
+
         return self.fd
 
     def flush(self, path, fh=None):
@@ -332,13 +469,16 @@ class CGFS(LoggingMixIn, Operations):
         parts = self.split_path(path)
         file = self.get_file(parts)
 
-        if file.stat is None and len(parts) > 3:
-            submission = self.get_submission(path)
+        if isinstance(file, TempFile):
+            return file.getattr()
 
+        if file.stat is None and len(parts) > 3:
             try:
-                query_path = submission.tld + '/' + '/'.join(parts[3:])
+                submission = self.get_submission(path)
             except CGAPIException as e:
                 handle_cgapi_exception(e)
+
+            query_path = submission.tld + '/' + '/'.join(parts[3:])
 
             if isinstance(file, Directory):
                 query_path += '/'
@@ -346,7 +486,10 @@ class CGFS(LoggingMixIn, Operations):
             submission = None
             query_path = None
 
-        return file.getattr(submission, query_path)
+        attrs = file.getattr(submission, query_path)
+        if self.fixed and isinstance(file, File):
+            attrs['st_mode'] &= ~0o222
+        return attrs
 
     # TODO?: Add xattr support
     def getxattr(self, path, name, position=0):
@@ -365,11 +508,14 @@ class CGFS(LoggingMixIn, Operations):
         if dname in parent.children:  # pragma: no cover
             raise FuseOSError(EEXIST)
 
-        submission = self.get_submission(path)
-        query_path = submission.tld + '/' + '/'.join(parts[3:]) + '/'
-        ddata = cgapi.create_file(submission.id, query_path)
+        if self.fixed:
+            parent.insert(TempDirectory({}, name=dname, writable=True))
+        else:
+            submission = self.get_submission(path)
+            query_path = submission.tld + '/' + '/'.join(parts[3:]) + '/'
+            ddata = cgapi.create_file(submission.id, query_path)
 
-        parent.insert(Directory(ddata, name=dname, writable=True))
+            parent.insert(Directory(ddata, name=dname, writable=True))
 
     def open(self, path, flags):
         parts = self.split_path(path)
@@ -388,23 +534,31 @@ class CGFS(LoggingMixIn, Operations):
             if flags & (O_CREAT & O_EXCL):  # pragma: no cover
                 raise FuseOSError(EEXIST)
 
-        if file.data is None:
-            file.open(cgapi.get_file(file.id))
+        if isinstance(file, TempFile):
+            file.open()
 
         if flags & O_TRUNC:
             file.truncate(0)
 
         self.fd += 1
+        self._open_files[self.fd] = file
         return self.fd
 
     def read(self, path, size, offset, fh):
-        file = self.get_file(path, expect_type=File)
+        if fh is not None and fh in self._open_files:
+            file = self._open_files[fh]
+        else:
+            file = self.get_file(path, expect_type=File)
+
+        if isinstance(file, TempFile):
+            return file.read(offset, size)
+
         return file.data[offset:offset + size]
 
     def readdir(self, path, fh):
         dir = self.get_dir(path)
 
-        if len(dir.children) == 0:
+        if not dir.children:
             if dir.type == DirTypes.ASSIGNMENT:
                 self.load_submissions(dir)
             elif dir.type == DirTypes.SUBMISSION:
@@ -416,6 +570,12 @@ class CGFS(LoggingMixIn, Operations):
         raise FuseOSError(EINVAL)
 
     def release(self, path, fh=None):
+        if fh is not None and fh in self._open_files:
+            file = self._open_files[fh]
+            file.release()
+            del self._open_files[fh]
+            return
+
         parts = self.split_path(path)
         parent = self.get_dir(parts[:-1])
         fname = parts[-1]
@@ -444,14 +604,17 @@ class CGFS(LoggingMixIn, Operations):
             raise FuseOSError(EPERM)
 
         new_query_path = submission.tld + '/' + '/'.join(new_parts[3:]) + '/'
-        print(new_query_path)
 
-        try:
-            res = cgapi.rename_file(file.id, new_query_path)
-        except CGAPIException as e:
-            handle_cgapi_exception(e)
+        if not isinstance(file, (TempDirectory, TempFile)):
+            if self.fixed:
+                raise FuseOSError(EPERM)
 
-        file.id = res['id']
+            try:
+                res = cgapi.rename_file(file.id, new_query_path)
+            except CGAPIException as e:
+                handle_cgapi_exception(e)
+
+            file.id = res['id']
         file.name = new_parts[-1]
 
         old_parent.pop(old_parts[-1])
@@ -464,13 +627,17 @@ class CGFS(LoggingMixIn, Operations):
 
         if dir.type != DirTypes.REGDIR:
             raise FuseOSError(EPERM)
-        if len(dir.children) != 0:
+        if dir.children:
             raise FuseOSError(ENOTEMPTY)
 
-        try:
-            cgapi.delete_file(dir.id)
-        except CGAPIException as e:
-            handle_cgapi_exception(e)
+        if not isinstance(dir, TempDirectory):
+            if self.fixed:
+                raise FuseOSError(EPERM)
+
+            try:
+                cgapi.delete_file(dir.id)
+            except CGAPIException as e:
+                handle_cgapi_exception(e)
 
         parent.pop(parts[-1])
 
@@ -494,7 +661,14 @@ class CGFS(LoggingMixIn, Operations):
         if length < 0:  # pragma: no cover
             raise FuseOSError(EINVAL)
 
-        file = self.get_file(path, expect_type=File)
+        if fh is not None and fh in self._open_files:
+            file = self._open_files[fh]
+        else:
+            file = self.get_file(path, expect_type=File)
+
+        if self.fixed and not isinstance(file, TempFile):
+            raise FuseOSError(EPERM)
+
         file.truncate(length)
 
     def unlink(self, path):
@@ -503,10 +677,16 @@ class CGFS(LoggingMixIn, Operations):
         fname = parts[-1]
         file = self.get_file(fname, start=parent, expect_type=File)
 
-        try:
-            cgapi.delete_file(file.id)
-        except CGAPIException as e:
-            handle_cgapi_exception(e)
+        if isinstance(file, TempFile):
+            file.unlink()
+        else:
+            if self.fixed:
+                raise FuseOSError(EPERM)
+
+            try:
+                cgapi.delete_file(file.id)
+            except CGAPIException as e:
+                handle_cgapi_exception(e)
 
         parent.pop(fname)
 
@@ -514,15 +694,25 @@ class CGFS(LoggingMixIn, Operations):
         file = self.get_file(path)
 
         atime, mtime = times or (time(), time())
-        file.setattr('st_atime', atime)
-        file.setattr('st_mtime', mtime)
+        if isinstance(file, TempFile):
+            file.utimens(atime, mtime)
+        else:
+            file.setattr('st_atime', atime)
+            file.setattr('st_mtime', mtime)
 
     def write(self, path, data, offset, fh):
-        file = self.get_file(path, expect_type=File)
+        if fh is not None and fh in self._open_files:
+            file = self._open_files[fh]
+        else:
+            file = self.get_file(path, expect_type=File)
+        if self.fixed and not isinstance(file, TempFile):
+            raise FuseOSError(EPERM)
         return file.write(data, offset)
 
 
 if __name__ == '__main__':
+    print('Mounting... ')
+
     argparser = ArgumentParser(description='CodeGra.de file system')
     argparser.add_argument(
         'username',
@@ -542,8 +732,18 @@ if __name__ == '__main__':
         metavar='PASSWORD',
         type=str,
         dest='password',
-        help='Your CodeGra.de password, don\' pass this option if you want to '
-        ' pass your password over stdin.'
+        help="""Your CodeGra.de password, don' pass this option if you want to
+        pass your password over stdin."""
+    )
+    argparser.add_argument(
+        '-u',
+        '--url',
+        metavar='URL',
+        type=str,
+        dest='url',
+        help="""The url to find the api. This defaults to
+        'https://codegra.de/api/v1/'. It can also be passed as a environment
+        variable 'CGAPI_BASE_URL'"""
     )
     argparser.add_argument(
         '-v',
@@ -561,16 +761,35 @@ if __name__ == '__main__':
         default=False,
         help='Only see the latest submissions of students.'
     )
+    argparser.add_argument(
+        '-f',
+        '--fixed',
+        dest='fixed',
+        action='store_true',
+        default=False,
+        help="""Mount the original files as read only. It is still possible to
+        create new files, but it is not possible to alter existing files."""
+    )
     args = argparser.parse_args()
 
     mountpoint = args.mountpoint
     username = args.username
     password = args.password if args.password is not None else getpass()
     latest_only = args.latest_only
+    fixed = args.fixed
 
     if args.debug:
         logging.basicConfig(level=logging.DEBUG)
 
-    cgapi = CGAPI(username, password)
+    cgapi = CGAPI(
+        username, password, args.url or getenv('CGAPI_BASE_URL', None)
+    )
 
-    fuse = FUSE(CGFS(latest_only), mountpoint, nothreads=True, foreground=True)
+    with tempfile.TemporaryDirectory(dir=tempfile.gettempdir()) as tmpdir:
+        print(tmpdir)
+        fuse = FUSE(
+            CGFS(latest_only, fixed=fixed, tmpdir=tmpdir),
+            mountpoint,
+            nothreads=True,
+            foreground=True
+        )
